requirements.txt
python-telegram-bot==21.5
gspread==6.1.4
google-auth==2.22.0
aiohttp==3.9.5
